ADD : addComment 컴포넌트 분리, icon.js 분리
