+++ src/pages/namsunkim/Feed/Components/CommentList/CommentList.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variable.scss';
+@import '../../../styles/variable.scss';
 
 .commentList{
   padding: 0 1rem 0.75rem 1rem;
