+++ src/pages/namsunkim/Feed/Components/FeedIcon/FeedIcon.scss
